Moving RTThreadBuffer to buffers.py
